Allow to add gt synapses during mapping
To take care of dandling nodes that come from synapses (and that should always be used for mapping): Before, the global number of skeleton nodes were used to estimate whether this is a dangling node, but this had some failure cases (eg. if there are multiple dangling nodes from the same skeleton, but in different segments). Now instead, do not ignore nodes, if they come from synapses.